Move rate limiter module to `experimental.evals/models`
