+++ tests/utilities/test_ratelimits.py
@@ -6,7 +6,7 @@ from typing import Optional
 from unittest import mock
 
 import pytest
-from phoenix.utilities.ratelimits import LimitStore, TokenRateLimiter, UnavailableTokensError
+from phoenix.experimental.evals.models.rate_limiters import LimitStore, TokenRateLimiter, UnavailableTokensError
 
 
 @contextmanager
